Clean up stories
